requirements.txt
# requirements.txt
aiogram
fastapi
uvicorn[standard]
python-dotenv
httpx
stripe
openai
apscheduler
matplotlib
pandas
mplfinance
pycoingecko
